added admin controler and move the getApplicants and getApplicantsState
