+++ server/controllers/admin/getApplicantsStats.js
