+++ server/controllers/admin/index.js
@@ -1,0 +1,7 @@
+const getApplicants = require('./getApplicants');
+const getApplicantsStats = require('./getApplicantsStats');
+
+module.exports = {
+  getApplicants,
+  getApplicantsStats,
+};
