+++ server/controllers/index.js
@@ -1,4 +1,5 @@
 const applicant = require('./applicant');
 const cohorts = require('./cohort/getCohorts');
+const admin = require('./admin');
 
-module.exports = { cohorts, applicant };
+module.exports = { cohorts, applicant, admin };
