+++ server/router/admin-router.js
@@ -2,10 +2,7 @@ const express = require('express');
 
 const adminRouter = express.Router();
 
-const {
-  getApplicantsStats,
-  getApplicants,
-} = require('../controllers/applicant');
+const { getApplicantsStats, getApplicants } = require('../controllers/admin');
 
 adminRouter.get('/applicants/stats', getApplicantsStats);
 adminRouter.get('/applicants', getApplicants);
